+++ libraries/API/include/api/api.hpp
@@ -9,6 +9,7 @@
 
 #include <memory>
 #include <vector>
+#include <functional>
 
 #include "macros.hpp"